+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add `:item_page` option to `Rapa::Client#search_items`
 - Add `Rapa::Queries::BaseQuery#inspect`
 
 ## 0.4.1